Refactored query params in a new query directive
